+++ src/index.ts
@@ -1,3 +1,16 @@
-import { createHandler } from "./handlers.js";
+import { createHandler, createHandlerFromSchema } from "./handlers.js";
+import {
+  mockEnums,
+  mockCustomScalars,
+  createPossibleTypesMap,
+  createDefaultResolvers,
+} from "./utilities.ts";
 
-export { createHandler };
+export {
+  createHandler,
+  createHandlerFromSchema,
+  mockEnums,
+  mockCustomScalars,
+  createPossibleTypesMap,
+  createDefaultResolvers,
+};
